fqdn: do not `TriggerPolicyUpdates`
Instead, update the `PolicyMap` directly for each endpoint. When we
update the selectorcache with the new mappings of `FQDNSelector` --> IP,
the PolicyMap updates are already prepared via `AccumulateMapChanges`
further down the call stack. This means that the PolicyMap is ready to be
updated already with the identities relating to IPs for the selectors
which were updated. We can do this out-of-bound of Endpoint regeneration
because no state related to policy (e.g., the rules) have changed as part
of this updating of the aforementioned mapping.

Signed-off by: Ian Vernon <ian@cilium.io>